Some cleanup to the recipe form page
