+++ extension/content/components/recipes/form/RecipeForm.js
@@ -1,6 +1,5 @@
 import React from "react";
-import { useHistory, useParams } from "react-router-dom";
-import { Alert, Button, ButtonToolbar, Form } from "rsuite";
+import { Form } from "rsuite";
 
 import GenericField from "devtools/components/recipes/form/fields/GenericField";
 import CodeMirrorField from "devtools/components/recipes/form/fields/CodeMirrorField";
@@ -8,56 +7,8 @@ import ActionPicker from "devtools/components/recipes/form/fields/ActionPicker";
 import ActionArguments from "devtools/components/recipes/form/ActionArguments";
 import FilterObjects from "devtools/components/recipes/form/FilterObjects";
 import ImportInstructions from "devtools/components/recipes/form/fields/ImportInstructions";
-import {
-  useEnvironmentState,
-  useSelectedNormandyEnvironmentAPI,
-} from "devtools/contexts/environment";
-import { useRecipeDetailsState } from "devtools/contexts/recipeDetails";
 
 export default function RecipeForm() {
-  const history = useHistory();
-  const { recipeId } = useParams();
-  const { selectedKey: environmentKey } = useEnvironmentState();
-  const { data, importInstructions } = useRecipeDetailsState();
-  const normandyApi = useSelectedNormandyEnvironmentAPI();
-
-  const handleSubmitClick = () => {
-    try {
-      if (importInstructions) {
-        throw Error("Import Instructions not empty!");
-      }
-
-      /* eslint-disable no-unused-vars */
-      const { comment: _omitComment, action, ...cleanedData } = data;
-      /* eslint-enable no-unused-vars */
-
-      const requestSave = normandyApi.saveRecipe(recipeId, {
-        ...cleanedData,
-        action_id: action.id,
-      });
-
-      requestSave
-        .then((savedRecipe) => {
-          history.push(`/${environmentKey}/recipes/${savedRecipe.id}`);
-          Alert.success("Changes Saved");
-        })
-        .catch((err) => {
-          console.warn(err.message, err.data);
-          Alert.error(`An Error Occurred: ${JSON.stringify(err.data)}`, 5000);
-        });
-    } catch (err) {
-      Alert.error(err.message);
-    }
-  };
-
-  const handleCancelClick = () => {
-    if (recipeId) {
-      history.push(`/${environmentKey}/recipes/${recipeId}`);
-    } else {
-      history.push(`/${environmentKey}/recipes`);
-    }
-  };
-
   return (
     <Form fluid>
       <GenericField required label="Name" name="name" />
@@ -76,15 +27,6 @@ export default function RecipeForm() {
       <ActionPicker />
 
       <ActionArguments />
-
-      <ButtonToolbar>
-        <Button appearance="primary" onClick={handleSubmitClick}>
-          Submit
-        </Button>
-        <Button appearance="default" onClick={handleCancelClick}>
-          Cancel
-        </Button>
-      </ButtonToolbar>
     </Form>
   );
 }
